+++ component/utils/loadAssets.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+
+export const bookOn = require('../../assets/icon/book_on.png');
+export const bookOff = require('../../assets/icon/book_off.png');
