--- 정리.js
@@ -1,50 +1,0 @@
-전체 변수
-let dataList = [];
-let informData = {};
-let pushList = [];
-let userData = [];
-const screenHeight, screenWidth;
-
-let pressDiaryEditButton = false;  // diary 편집버튼 누름 상태값
-let global_p_id = 0;               // 채팅창 사이드 메뉴에서 다른 상품정보로 보내기 위한 상품 id 값
-let editDiaryTextMode = false;     // 다이어리 편집모드 상태값
-let global_y = 0;         // 다이어리리스트 스크린의 스크롤 값
-
-독립 함수
-function chooseRandomIndex(a){}
-function chooseRandomly(a){}
-function diarySortByDate(myDiaryMessageList){}
-function isEmail(email){}
-
-
-PUSH TEST
-function pushMessage(id){
-  dataList
-  chooseRandomIndex()
-}
-function pushTestHandler(updateScreenHandler){
-  userData
-  chooseRandomly()
-  function pushMessage(id){
-}
-
-DIARY
-function diaryPosToRealPos(diaryPos){}
-function realPosToDiaryPos(realPos){
-  userData
-}
-function DraggableDiary({id, changePosHandler, nav, updateDiary, cancelDrag}){
-  userData
-  function diaryPosToRealPos(diaryPos){}
-  <AnimatableDiaryComponent>
-}
-function BasiceDiary({id, changePosHandler, nav}){
-  userData
-  function diaryPosToRealPos(diaryPos){}
-  <DiaryComponent id={id} nav={nav}/>
-}
-function AnimatableDiaryComponent(props){
-  dataList
-  userData
-  
-}
